Create README.md
This is an example of how Bike-Share riders (casual and member) differ. The data show that casual riders are more likely to ride for longer periods of time. More importantly, the average casual rider prefers riding on Fridays rather than most days of the week. Additionally, the data show that member riders and casual riders have a preference for starting location for riding.